microseconds should be unsigned long
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@7239 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/DemoCamera/DemoCamera.cpp
@@ -534,7 +534,8 @@ int CDemoCamera::SnapImage()
       long naptime = (long)(0.5 + expUs - (double)(t2-startTime).getUsec());
       if( naptime < 1)
          naptime = 1;
-      CDeviceUtils::NapMicros(naptime);
+      // longest possible nap is about 38 minutes
+      CDeviceUtils::NapMicros((unsigned long) naptime);
    }
    else
    {
